updating

+++ index.js
@@ -6,25 +6,40 @@ var fork = require('child_process').fork;
 
 var server;
 
-var HUB_DIR = process.env.HUB_DIR || '/tmp/hub';
+var hub = process.env.HUB === 'true';
 
-var REPO_URL = process.env.REPO_URL || 'https://github.com/serandules/hub.git';
+var dev = process.env.NODE_ENV === 'development';
 
-var start = function (dir, repo) {
-    rimraf.sync(dir); //TODO
-    fs.mkdirSync(dir);
+var HUB_DIR = process.env.HUB_DIR || '/tmp/';
+
+var HUB_REPO = 'https://github.com/serandules/hub.git';
+
+var HUBCLIENT_REPO = 'https://github.com/serandules/hub-client.git';
+
+var start = function () {
+    var dir = HUB_DIR;
+    console.log(dev);
+    if (!dev) {
+        rimraf.sync(dir); //TODO
+        fs.mkdirSync(dir);
+    }
     var child = spawn('bash');
-    var id = uuid.v4();
+    var repo = hub ? HUB_REPO : HUBCLIENT_REPO;
+    var id = dev ? (hub ? 'hub' : 'hub-client') : uuid.v4();
     child.stdout.pipe(process.stdout);
     child.stderr.pipe(process.stderr);
     child.stdin.write('pwd\n');
     child.stdin.write('cd ' + HUB_DIR + '\n');
     child.stdin.write('pwd\n');
-    console.log('git clone ' + repo + ' ' + id + '\n');
-    child.stdin.write('git clone ' + repo + ' ' + id + '\n');
+    if (!fs.existsSync(HUB_DIR + id)) {
+        console.log('git clone ' + repo + ' ' + id + '\n');
+        child.stdin.write('git clone ' + repo + ' ' + id + '\n');
+    }
     child.stdin.write('cd ' + id + '\n');
     child.stdin.write('npm install\n');
-    child.stdin.write('component install\n');
+    if (hub) {
+        child.stdin.write('component install\n');
+    }
     child.stdin.write('echo "repo cloned"\n');
     child.stdin.write('kill -9 ' + child.pid + '\n');
     child.on('close', function () {
@@ -33,7 +48,7 @@ var start = function (dir, repo) {
             server.kill('SIGKILL');
             console.log('old server killed');
         }
-        var cwd = dir + '/' + id;
+        var cwd = dir + id;
         server = fork(cwd + '/index.js', {
             cwd: cwd,
             silent: true
@@ -50,4 +65,4 @@ var start = function (dir, repo) {
     });
 };
 
-start(HUB_DIR, REPO_URL);
+start();
